Shifting Google Oauth details closer to LoginSignUp component
